Update EntityLocalInterface for UseCase handling
- Add API to remove multiple usecases in a single step
- Remove API to only remove a single usecase
- Update UseCase API calls, to use a new model.UseCaseFilter struct, which is used to invoke the action on all usecases matching the filter

Reasoning: Every individal usecase removal will trigger a SPINE message, and if multiple are remove, this triggers multiple messages which is not a desired outcome even though it would be valid. So the intention is to get only a single SPINE notify message with all removals already being part of.