Add JSON Marshalling support to device and entity
Both DeviceInterface and EntityInterface implementations (e.g. DeviceRemoteInterface and EntityRemoteInterface) are used as arguments in API calls and return values.

When these API calls are used via interprocess communication protocols based on JSON, then the easiest way is to convert them into their addresses to be identified.

This change adds support for this need.